+++ lib/physics/aabb.dart
@@ -4,7 +4,6 @@ import 'dart:ui';
 import 'package:tremble/physics/circle.dart';
 import 'package:tremble/physics/shape.dart';
 import 'package:tremble/physics/vec2.dart';
-import 'package:tremble/utils/extensions.dart';
 
 class AABB extends Shape {
   double width;
@@ -70,12 +69,4 @@ class AABB extends Shape {
 
   @override
   AABB get aabb => this;
-
-  @override
-  void draw(Canvas canvas, Color color, [bool drawBoundingBox = true]) {
-    Shape.paint.style = PaintingStyle.fill;
-    Shape.paint.color = color;
-    canvas.drawRect(rect, Shape.paint);
-    if (drawBoundingBox) super.draw(canvas, color.inverted);
-  }
 }